Rearrange ColorPropagation
