+++ arangod/Pregel/Algos/ColorPropagation/ColorPropagation.cpp
