+++ arangod/Pregel/CommonFormats.h
@@ -82,83 +82,5 @@ struct ECValue {
   std::vector<uint32_t> shortestPaths;
 };
 
-using CollectionIdType = uint16_t;
-using ColorType = uint16_t;
-using PropagatedColor = uint16_t;
-using VectorOfColors = std::vector<PropagatedColor>;
-
-struct ColorPropagationValue {
-  CollectionIdType equivalenceClass;
-  std::vector<bool> colors;
-
-  static CollectionIdType none() {
-    return std::numeric_limits<CollectionIdType>::max();
-  }
-
-  [[nodiscard]] bool contains(ColorType color) const {
-    // todo assert that color < numColors
-    return colors[color];
-  }
-
-  void add(ColorType color) {
-    // todo assert that color < numColors
-    colors[color] = true;
-  }
-
-  [[nodiscard]] VectorOfColors getColors() const {
-    VectorOfColors result;
-    auto const size = colors.size();
-    result.reserve(size);
-    for (size_t color = 0; color < size; ++color) {
-      if (contains(static_cast<ColorType>(color))) {
-        result.push_back(static_cast<decltype(result)::value_type>(color));
-      }
-    }
-    result.shrink_to_fit();
-    return result;
-  }
-};
-
-struct ColorPropagationMessageValue {
-  CollectionIdType equivalenceClass = 0;
-  std::vector<PropagatedColor> colors;
-};
-
-template<typename Inspector>
-auto inspect(Inspector& f, ColorPropagationMessageValue& x) {
-  return f.object(x).fields(
-      f.field(Utils::equivalenceClass, x.equivalenceClass),
-      f.field(Utils::colors, x.colors));
-}
-
-struct ColorPropagationUserParameters {
-  uint64_t maxGss;
-  uint16_t numColors;
-  std::string inputColorsFieldName;
-  std::string outputColorsFieldName;
-  std::string equivalenceClassFieldName;
-};
-
-template<typename Inspector>
-auto inspect(Inspector& f, ColorPropagationUserParameters& x) {
-  return f.object(x).fields(
-      f.field(Utils::maxGSS, x.maxGss), f.field(Utils::numColors, x.numColors),
-      f.field(Utils::inputColorsFieldName, x.inputColorsFieldName),
-      f.field(Utils::outputColorsFieldName, x.outputColorsFieldName),
-      f.field(Utils::equivalenceClass, x.equivalenceClassFieldName));
-}
-
-struct ColorPropagationValueMessageFormat
-    : public MessageFormat<ColorPropagationMessageValue> {
-  ColorPropagationValueMessageFormat() = default;
-  void unwrapValue(VPackSlice s,
-                   ColorPropagationMessageValue& value) const override {
-    value = deserialize<ColorPropagationMessageValue>(s);
-  }
-  void addValue(VPackBuilder& arrayBuilder,
-                ColorPropagationMessageValue const& value) const override {
-    serialize(arrayBuilder, value);
-  }
-};
 
 }  // namespace arangodb::pregel
