+++ arangod/Pregel/IncomingCache.cpp
@@ -27,6 +27,8 @@
 #include "Pregel/Worker/Messages.h"
 #include "Pregel/Worker/WorkerConfig.h"
 
+#include "Pregel/Algos/ColorPropagation/ColorPropagationValue.h"
+
 #include "Basics/MutexLocker.h"
 #include "Basics/StaticStrings.h"
 #include "Basics/VelocyPackHelper.h"
