+++ arangod/Pregel/Worker/GraphStore.cpp
@@ -54,6 +54,7 @@
 #include "VocBase/ticks.h"
 #include "VocBase/vocbase.h"
 
+#include "Pregel/Algos/ColorPropagation/ColorPropagationValue.h"
 #include "Pregel/Algos/HITS/HITSValue.h"
 #include "Pregel/Algos/HITSKleinberg/HITSKleinbergValue.h"
 #include "Pregel/Algos/LabelPropagation/LPValue.h"
